requirements.txt
torch
torchvision
easydict
opencv_python
numpy
scipy
Pillow
